requirements.txt
streamlit
pandas
numpy
tensorflow
nltk
pillow
pytesseract